fix(backend/tests): Place prismaTestUtils outside of tests folder to avoid Jest errors
